Added draft gRPC server. Refactored the listener. Updated the proto interface.
