+++ cmd/nanotube/main.go
@@ -16,6 +16,7 @@ import (
 	"time"
 
 	"github.com/bookingcom/nanotube/pkg/conf"
+	"github.com/bookingcom/nanotube/pkg/in"
 	"github.com/bookingcom/nanotube/pkg/metrics"
 	"github.com/bookingcom/nanotube/pkg/rewrites"
 	"github.com/bookingcom/nanotube/pkg/rules"
@@ -100,7 +101,7 @@ func main() {
 
 	stop := make(chan struct{})
 	n := gracenet.Net{}
-	queue, err := Listen(&n, &cfg, stop, lg, ms)
+	queue, err := in.Listen(&n, &cfg, stop, lg, ms)
 	if err != nil {
 		log.Fatalf("error launching listener, %v", err)
 	}
@@ -116,6 +117,7 @@ func main() {
 	sgn := make(chan os.Signal, 1)
 	signal.Notify(sgn, os.Interrupt, syscall.SIGTERM, syscall.SIGUSR2)
 
+	// TODO: Make seamless restart work for GRPC
 	for {
 		s := <-sgn
