--- docker-compose.yaml
@@ -1,23 +1,0 @@
-version: "3"
-
-services:
-  go-carbon-0:
-    build:
-      dockerfile: docker/go-carbon/Dockerfile
-      context: .
-    ports:
-    - "8080:8080"
-
-  go-carbon-1:
-    build:
-      dockerfile: docker/go-carbon/Dockerfile
-      context: .
-    ports:
-    - "8081:8080"
-
-  nanotube:
-    build:
-      dockerfile: docker/nanotube/Dockerfile
-      context: .
-    ports:
-      - "2003:2003"
