--- docker/go-carbon/go-carbon.conf
@@ -1,85 +1,0 @@
-# See https://github.com/bookingcom/go-carbon for configuration details
-[common]
-user = ""
-graph-prefix = "carbon.agents.{host}"
-metric-endpoint = "local"
-metric-interval = "1m0s"
-max-cpu = 4
-[whisper]
-data-dir = "/data/graphite/whisper"
-schemas-file = "/etc/storage-schemas.conf"
-aggregation-file = "/etc/storage-aggregation.conf"
-workers = 8
-max-updates-per-second = 0
-sparse-create = false
-enabled = true
-[cache]
-max-size = 1000000
-write-strategy = "max"
-[udp]
-listen = ":2003"
-enabled = true
-log-incomplete = false
-buffer-size = 0
-[tcp]
-listen = ":2003"
-enabled = true
-buffer-size = 0
-[pickle]
-listen = ":2004"
-max-message-size = 67108864
-enabled = true
-buffer-size = 0
-[carbonlink]
-listen = ":7002"
-enabled = true
-read-timeout = "30s"
-[grpc]
-listen = ":7003"
-enabled = true
-[carbonserver]
-listen = ":8080"
-enabled = true
-buckets = 10
-metrics-as-counters = false
-read-timeout = "60s"
-write-timeout = "60s"
-query-cache-enabled = true
-query-cache-size-mb = 0
-find-cache-enabled = false #set to true
-trigram-index = true
-postings-index = false
-scan-frequency = "5m0s"
-max-globs = 100
-fail-on-max-globs = false
-graphite-web-10-strict-mode = true
-internal-stats-dir = ""
-stats-percentiles = [99, 98, 95, 75, 50]
-[dump]
-enabled = false
-path = "/var/lib/graphite/dump/"
-restore-per-second = 0
-[pprof]
-listen = "localhost:7007"
-enabled = true
-[[logging]]
-logger = ""
-file= "stdout"
-level = "info"
-encoding = "json"
-encoding-time = "iso8601"
-encoding-duration = "seconds"
-[[logging]]
-logger = "stat"
-file= "none"
-level = "info"
-encoding = "json"
-encoding-time = "iso8601"
-encoding-duration = "seconds"
-[[logging]]
-logger = "carbonserver"
-file= "stdout"
-level = "debug"
-encoding = "json"
-encoding-time = "iso8601"
-encoding-duration = "seconds"
