--- docker/go-carbon/storage-aggregation.conf
@@ -1,7 +1,0 @@
-# Documentation:
-# http://graphite.readthedocs.io/en/latest/config-carbon.html#storage-aggregation-conf
-
-[default]
-pattern = .*
-xFilesFactor = 0.5
-aggregationMethod = average
