--- docker/go-carbon/storage-schemas.conf
@@ -1,6 +1,0 @@
-# Documentation:
-# http://graphite.readthedocs.io/en/latest/config-carbon.html#storage-schemas-conf
-
-[default]
-pattern = .*
-retentions = 60s:30d,1h:5y
