--- docker/nanotube/config/clusters.toml
@@ -1,15 +1,0 @@
-[[cluster]]
-name = "carbon0"
-type = "jump"
-    [[cluster.hosts]]
-    name = "go-carbon-0"
-    port = 2003
-    index = 0
-
-[[cluster]]
-name = "carbon1"
-type = "jump"
-    [[cluster.hosts]]
-    name = "go-carbon-1"
-    port = 2003
-    index = 0
