--- docker/nanotube/config/rules.toml
@@ -1,16 +1,0 @@
-[[rule]]
-    regexs = [
-        ".*test1.*"
-    ]
-    clusters = [
-        "carbon0"
-    ]
-    continue = true
-
-[[rule]]
-    regexs = [
-        ".*test2.*"
-    ]
-    clusters = [
-        "carbon1"
-    ]
