+++ pkg/in/grpc.go
@@ -1,0 +1,79 @@
+package in
+
+import (
+	"io"
+	"net"
+	"sync"
+
+	"github.com/bookingcom/nanotube/pkg/grpcstreamer"
+	"github.com/bookingcom/nanotube/pkg/metrics"
+	"go.uber.org/zap"
+	"google.golang.org/grpc"
+)
+
+// TODO: Add HTTP tracing https://blog.golang.org/http-tracing
+
+// Blocking. Returns when server returns.
+func listenGRPC(l net.Listener, queue chan string, stop <-chan struct{}, connWG *sync.WaitGroup, ms *metrics.Prom, lg *zap.Logger) {
+	// TODO: Check for optimal server options
+	s := streamerServer{
+		queue: queue,
+		stop:  stop,
+
+		lg: lg,
+		ms: ms,
+	}
+	gRPCServer := grpc.NewServer()
+	grpcstreamer.RegisterStreamerServer(gRPCServer, &s)
+	err := gRPCServer.Serve(l)
+	if err != nil {
+		lg.Error("serving gRPC failed", zap.Error(err))
+	}
+
+	connWG.Done()
+}
+
+type streamerServer struct {
+	queue chan string
+	stop  <-chan struct{}
+
+	lg *zap.Logger
+	ms *metrics.Prom
+}
+
+// This function can be running in multiple goroutines simultaneously. Each new incoming
+// transmission will make a new instance.
+func (server *streamerServer) Stream(s grpcstreamer.Streamer_StreamServer) error {
+	res := grpcstreamer.Result{}
+
+loop:
+	for {
+		select {
+		case <-server.stop:
+			err := s.SendAndClose(&res)
+			if err != nil {
+				server.lg.Error("error when sending a response to stream while stopping", zap.Error(err))
+			}
+			break loop
+		default:
+		}
+		rec, err := s.Recv()
+
+		if err == io.EOF {
+			// TODO: Do we have to do something special to re-start listening?
+			err := s.SendAndClose(&res)
+			if err != nil {
+				server.lg.Error("error when sending a response to stream", zap.Error(err))
+			}
+		}
+		if err != nil {
+			res.ErrorCount++
+			// TODO: Is it ok to continue after an error?
+			continue
+		}
+		res.ReceivedCount++
+		server.queue <- string(rec.Rec)
+	}
+
+	return nil
+}
