--- pkg/linegrpc/line.proto
@@ -1,17 +1,0 @@
-syntax = "proto3";
-
-option go_package = "linegrpc";
-
-service Main {
-    rpc Send(stream Rec) returns (Summary) {}
-}
-
-message Rec {
-    string path = 1;
-    double val = 2;
-    uint32 time = 3;
-}
-
-message Summary {
-    uint32 receivedCount = 1;
-}
